+++ src/scene/entity/ComponentClass.js
@@ -19,7 +19,11 @@ export class ComponentClass {
    * @param {ComponentNewCallback<T>} [newCallback]
    * @param {ComponentDeleteCallback<T>} [deleteCallback]
    */
-  constructor(name, newCallback = () => /** @type {T} */ (null), deleteCallback = () => {}) {
+  constructor(
+    name,
+    newCallback = () => /** @type {T} */ (null),
+    deleteCallback = () => {},
+  ) {
     /** @type {string} */
     this.name = name;
     /** @type {ComponentNewCallback<T>} */