+++ src/scene/entity2/index.js
@@ -1,0 +1,5 @@
+export * from './EntityManager';
+export * from './QueryManager';
+export * from './Query';
+export * from './ComponentFactory';
+export * from './Archetype';