+++ src/scene/index.js
@@ -1,5 +1,6 @@
 export * from './camera/index.js';
 export * from './scenegraph/index.js';
-export * from './entity/index.js';
+export * as EntityV1 from './entity/index.js';
 export * from './topic/index.js';
 export * from './loop/index.js';
+export * from './entity2/index.js';